Moved files to symmetrical layout
library at top level and command lines under cmd/
(Assuming go get will work, not sure how to test that in a branch?)
If we split into multiple packages we can move it under
fortio/pkg/fortio but that seems ugly right now [2x fortio in the path,
though it’s there for cmd/fortio/…])
