+++ devel/fortio/BUILD.bazel
@@ -1,8 +1,21 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
-go_binary(
-    name = "fortio",
-    srcs = ["main.go"],
-    deps = ["//devel/fortio/fortioLib:go_default_library"],
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "http.go",
+        "periodic.go",
+        "stats.go",
+    ],
     visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "periodic_test.go",
+        "stats_test.go",
+    ],
+    size = "small",
+    library = ":go_default_library",
 )
