+++ devel/fortio/cmd/echosrv/BUILD.bazel
@@ -1,0 +1,7 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary")
+
+go_binary(
+    name = "echosrv",
+    srcs = ["echo.go"],
+    visibility = ["//visibility:public"],
+)
