+++ devel/fortio/cmd/fortio/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary")
+
+go_binary(
+    name = "fortio",
+    srcs = ["main.go"],
+    deps = ["//devel/fortio:go_default_library"],
+    visibility = ["//visibility:public"],
+)
