+++ devel/fortio/cmd/histogram/BUILD.bazel
@@ -4,7 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = ["main.go"],
     visibility = ["//visibility:private"],
-    deps = ["//devel/fortio/fortioLib:go_default_library"],
+    deps = ["//devel/fortio:go_default_library"],
 )
 
 go_binary(
