+++ devel/fortio/stats.go
